refactor: rename learner apis to learner api
